# Backend/docs/swagger.json
{"openapi":"3.0.0","info":{"title":"Inventory Management API","version":"1.0.0","description":"API docs for FiMoney Inventory backend"},"servers":[{"url":"http://localhost:8080"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"security":[{"bearerAuth":[]}],"paths":{"/products":{"post":{"summary":"Add a new product","tags":["Products"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name","type","sku","quantity","price"],"properties":{"name":{"type":"string"},"type":{"type":"string"},"sku":{"type":"string"},"image_url":{"type":"string"},"description":{"type":"string"},"quantity":{"type":"integer"},"price":{"type":"number"}}}}}},"responses":{"201":{"description":"Product created"},"400":{"description":"Missing required fields"},"500":{"description":"Server Error"}}},"get":{"summary":"Get all products for the logged-in user","tags":["Products"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"page","schema":{"type":"integer"},"description":"Page number for pagination"},{"in":"query","name":"limit","schema":{"type":"integer"},"description":"Number of items per page"}],"responses":{"200":{"description":"List of products"},"500":{"description":"Server Error"}}}},"/products/{id}/quantity":{"put":{"summary":"Update the quantity of a product","tags":["Products"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"The product ID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"quantity":{"type":"integer"}}}}}},"responses":{"200":{"description":"Product updated"},"400":{"description":"Quantity is required"},"404":{"description":"Product not found or not owned by user"},"500":{"description":"Server Error"}}}},"/register":{"post":{"summary":"Register a new user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"201":{"description":"User registered successfully"},"409":{"description":"User already exists"},"500":{"description":"Server Error"}}}},"/login":{"post":{"summary":"Login an existing user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}},"responses":{"200":{"description":"Login successful, returns JWT token"},"401":{"description":"Invalid credentials"},"500":{"description":"Server Error"}}}},"/admin/users":{"get":{"summary":"Get all users (Admin only)","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"List of all users (excluding passwords)","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"username":{"type":"string"},"isAdmin":{"type":"boolean"}}}}}}},"401":{"description":"Unauthorized - Missing or invalid token"},"403":{"description":"Forbidden - Not an admin"},"500":{"description":"Server error while fetching users"}}}},"/admin/analytics":{"get":{"summary":"Get basic product analytics (Admin only)","tags":["Admin"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Analytics data fetched successfully","content":{"application/json":{"schema":{"type":"object","properties":{"totalProducts":{"type":"integer"},"mostAddedSKU":{"type":"string"}}}}}},"401":{"description":"Unauthorized - Token missing or invalid"},"403":{"description":"Forbidden - Only admin can access"},"500":{"description":"Server Error"}}}}},"tags":[]}